requirements.txt
# Core dependencies
streamlit>=1.28.0
Pillow>=9.5.0
easyocr>=1.7.0
googletrans==4.0.0-rc1
PyMuPDF>=1.23.0

# PDF generation
reportlab>=4.0.0

# Additional utilities
numpy>=1.24.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0

# Optional: For better OCR performance
# paddlepaddle>=2.5.0
# paddleocr>=2.7.0
